requirements.txt
Pillow
emoji
aiohttp
aiofiles